+++ test/case/atbus_node_reg_test.cpp
@@ -17,8 +17,7 @@
 #endif
 
 #include <atbus_node.h>
-
-#include "detail/libatbus_protocol.h"
+#include <libatbus_protocol.h>
 
 #include "atbus_test_utils.h"
 #include "frame/test_macros.h"